Set arcana with nickname, fusion cancel fix
You can now set an arcana by using the nickname. If an arcana has a class which is not an arcana, but its nickname exists as an arcana name then it consider's the nickname.
Fixed a bug when a fusion was canceled and another persona was chosen (even with a wrong combination) it asked if you wanted to create the previous arcana.