Persona release additions and persona release, menu and fusion bug fixes
Personas window now correctly does not take any input if there are no personas.
Fixed empty line in personas window after releasing the last persona on the list (selects first persona).
Added additional messages to message window when release a persona (if persona is equipped or if it is the actor's last persona).
Correctly handles unordered persona fusion.
Personas that are already in the party have now disabled color in the results window.
Minor bug fixes.